Fixes all forms to use html form validation
When the submit buttons are pressed on any form, checkValidity is
called on the form to check if all the inputs are valid BEFORE calling
calling preventDefault on the submit event.

This allows the html form validation to alert the user when any inputs
are invalid before cancelling out the default behavior of form
submission.