Implement editing todo items
An edit button was added to the details section of each todo item. An
editing form replaces the normal display for each todo item that allows
the user to input values for each property of the todo item to alter
the todo items on submission.

There is a bug that needs to be fixed where the edit form submit button
refreshes the page if the user tries to submit an invalid due date
before submitting a valid due date.